fix vs warnings and treat warnings as errors from now on

+++ source/tracking/algorithm/fishparticle/particlefilter/ParticleClusters.cpp
@@ -18,7 +18,7 @@ ParticleClusters::~ParticleClusters(void)
 
 void ParticleClusters::cluster(const std::vector<Particle>& particles, unsigned num_clusters) {
 	// Prepare arguments for cv::kmeans.
-	cv::Mat_<float> data(particles.size(), 2);
+	cv::Mat_<float> data(static_cast<int>(particles.size()), 2);
 	for (int i = 0; i < data.rows; i++) {
 		data[i][0] = particles[i].getX();
 		data[i][1] = particles[i].getY();
